+++ src/timer.h
@@ -13,7 +13,7 @@ public:
 	void relaunch();
 	bool isActive() const;
 	unsigned getTicks() const;
-	unsigned getProgress() const;
+	double getProgress() const;
 	void setTicks(unsigned int ticks);
 	void stop();
 	void adjust(unsigned int newMaximum);
